ShoppyGO/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-jazzmin @ git+https://github.com/farridav/django-jazzmin.git@d269874c96fe5ff6d231c3017cdd4650bc2f3105
idna==3.10
pillow==11.2.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
stripe==12.0.0
typing_extensions==4.13.2
urllib3==2.4.0
python-dotenv==1.0.1
